+++ Databases/initialize_monster_database.py
@@ -40,10 +40,10 @@ def main():
                                     monster text PRIMARY KEY,
                                     hit_points integer,
                                     attack_speed integer,
-                                    chance_to_hit float,
+                                    chance_to_hit integer,
                                     minimum_damage integer,
                                     maximum_damage integer,
-                                    chance_to_heal float,
+                                    chance_to_heal integer,
                                     minimum_heal_points integer,
                                     maximum_heal_points integer
